refactor(nemesis): extract Scylla Manager nemesis into monkey/manager.py
Phase 4 of the nemesis extraction plan (nemesis-extraction.md).

Move all Manager-related disruption logic out of the monolithic
NemesisRunner and monkey/__init__.py into dedicated modules:

- Create sdcm/nemesis/monkey/manager.py with five self-contained
  monkey classes (MgmtBackup, MgmtBackupSpecificKeyspaces,
  MgmtRestore, MgmtRepair, MgmtCorruptThenRepair) and their
  module-level helper functions.

- Create sdcm/nemesis/utils/common_ops.py with
  destroy_data_and_restart_scylla() as a standalone function,
  extracted from NemesisRunner._destroy_data_and_restart_scylla.
  This is a cross-group utility also needed by data-destroy and
  streaming nemesis in later phases.

- Update ManagerRcloneBackup and ManagerNativeBackup in runners.py
  to call the new manager_backup() function directly instead of
  self.disrupt_manager_backup().

- Leave a thin wrapper _destroy_data_and_restart_scylla() on
  NemesisRunner so existing callers continue to work unchanged.

Removes ~350 lines from nemesis/__init__.py and ~60 lines from
monkey/__init__.py.

Refs: SCT-210
