+++ sdcm/nemesis/monkey/__init__.py
@@ -4,6 +4,13 @@ Classes can be used in nemesis_selector
 """
 
 from sdcm.nemesis import target_all_nodes, NemesisBaseClass, target_data_nodes
+from sdcm.nemesis.monkey.manager import (  # noqa: F401 — re-export for backward compatibility
+    MgmtBackup,
+    MgmtBackupSpecificKeyspaces,
+    MgmtCorruptThenRepair,
+    MgmtRepair,
+    MgmtRestore,
+)
 
 
 @target_all_nodes
@@ -343,65 +350,6 @@ class ToggleGcModeMonkey(NemesisBaseClass):
 
     def disrupt(self):
         self.runner.disrupt_toggle_table_gc_mode()
-
-
-@target_data_nodes
-class MgmtBackup(NemesisBaseClass):
-    manager_operation = True
-    disruptive = False
-    limited = True
-    supports_high_disk_utilization = False  # Snapshot/Restore operations consume disk space
-
-    def disrupt(self):
-        self.runner.disrupt_mgmt_backup()
-
-
-@target_data_nodes
-class MgmtBackupSpecificKeyspaces(NemesisBaseClass):
-    manager_operation = True
-    disruptive = False
-    limited = True
-    supports_high_disk_utilization = False  # Snapshot/Restore operations consume disk space
-
-    def disrupt(self):
-        self.runner.disrupt_mgmt_backup_specific_keyspaces()
-
-
-@target_data_nodes
-class MgmtRestore(NemesisBaseClass):
-    manager_operation = True
-    disruptive = True
-    kubernetes = True
-    xcloud = True
-    limited = True
-    supports_high_disk_utilization = False  # Snapshot/Restore operations consume disk space
-
-    def disrupt(self):
-        self.runner.disrupt_mgmt_restore()
-
-
-@target_data_nodes
-class MgmtRepair(NemesisBaseClass):
-    manager_operation = True
-    disruptive = False
-    kubernetes = True
-    limited = True
-
-    def disrupt(self):
-        self.runner.log.info("disrupt_mgmt_repair_cli Nemesis begin")
-        self.runner.disrupt_mgmt_repair_cli()
-        self.runner.log.info("disrupt_mgmt_repair_cli Nemesis end")
-        # For Manager APIs test, use: self.runner.disrupt_mgmt_repair_api()
-
-
-@target_data_nodes
-class MgmtCorruptThenRepair(NemesisBaseClass):
-    manager_operation = True
-    disruptive = True
-    kubernetes = True
-
-    def disrupt(self):
-        self.runner.disrupt_mgmt_corrupt_then_repair()
 
 
 class AbortRepairMonkey(NemesisBaseClass):
